#269 - disbaled deletion of default categories; styled list view (circles, etc.); sorted categories by name

+++ src/main/java/de/deadlocker8/budgetmaster/controller/CategoryController.java
@@ -1,5 +1,7 @@
 package de.deadlocker8.budgetmaster.controller;
 
+import de.deadlocker8.budgetmaster.entities.Category;
+import de.deadlocker8.budgetmaster.entities.CategoryType;
 import de.deadlocker8.budgetmaster.repositories.CategoryRepository;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Controller;
@@ -18,25 +20,38 @@ public class CategoryController extends BaseController
 	@RequestMapping("/categories")
 	public String index(@ModelAttribute("model") ModelMap model)
 	{
-		model.addAttribute("categories", categoryRepository.findAll());
+		model.addAttribute("categories", categoryRepository.findAllByOrderByNameAsc());
 		return "categories";
 	}
 
 	@RequestMapping("/categories/{ID}/requestDelete")
 	public String requestDeleteCategory(@ModelAttribute("model") ModelMap model, @PathVariable("ID") Integer ID)
 	{
+		if(!isDeletable(ID))
+		{
+			return "redirect:/categories";
+		}
+
 		model.addAttribute("confirm", true);
 		model.addAttribute("categories", categoryRepository.findAll());
 		model.addAttribute("currentCategory", categoryRepository.getOne(ID));
 		return "categories";
-
-		//TODO: default categories are not allowed for deletion --> deny (also hide delete buttons in html)
 	}
 
 	@RequestMapping("/categories/{ID}/delete")
 	public String deleteCategory(@ModelAttribute("model") ModelMap model, @PathVariable("ID") Integer ID)
 	{
-		categoryRepository.delete(ID);
+		if(isDeletable(ID))
+		{
+			categoryRepository.delete(ID);
+		}
+
 		return "redirect:/categories";
+	}
+
+	private boolean isDeletable(Integer ID)
+	{
+		Category categoryToDelete = categoryRepository.getOne(ID);
+		return categoryToDelete != null && categoryToDelete.getType() == CategoryType.CUSTOM;
 	}
 }
